fix audit allow delete cascade

+++ server/src/use-case/concert/delete-concert.ts
@@ -1,15 +1,30 @@
 import { Injectable } from '@nestjs/common';
+import { isEmpty } from 'lodash';
 
-import { IConcertRepository } from '@src/core/repository';
+import {
+  IConcertRepository,
+  IReservationRepository,
+} from '@src/core/repository';
 import { ConcertEntity } from '@src/core/entity';
 import { AppException } from '@src/common';
 
 @Injectable()
 export class DeleteConcertUseCase {
-  constructor(private concertRepo: IConcertRepository) {}
+  constructor(
+    private concertRepo: IConcertRepository,
+    private reservationRepository: IReservationRepository,
+  ) {}
 
   async findConcert(concertId: number): Promise<ConcertEntity> {
     return this.concertRepo.findById(concertId);
+  }
+
+  async canDeleteConcert(concertId: number): Promise<boolean> {
+    const tx = await this.reservationRepository.getTx();
+    const reserves = await tx.findMany({
+      where: { concertId },
+    });
+    return isEmpty(reserves);
   }
 
   async delete(concertId: number): Promise<ConcertEntity> {
@@ -17,6 +32,10 @@ export class DeleteConcertUseCase {
 
     if (!concert) {
       throw AppException.concertNotFound();
+    }
+
+    if (!(await this.canDeleteConcert(concertId))) {
+      throw AppException.cannotDeleteConcert();
     }
 
     return this.concertRepo.deleteById(concertId);
